--- project1_task-management/solution/optimizer/__init__.py
@@ -1,3 +1,0 @@
-from .creds import license_params
-
-__all__ = ["license_params"]
